move getRecommended function to utils
